Refactor orm.py to support pytest tests
