+++ python_model_service/orm/guid.py
@@ -1,0 +1,40 @@
+import uuid
+
+from sqlalchemy import TypeDecorator, CHAR
+from sqlalchemy.dialects.postgresql import UUID
+
+
+class GUID(TypeDecorator):  # pylint: disable=abstract-method
+    """Platform-independent GUID type.
+
+    Uses Postgresql's UUID type, otherwise uses
+    CHAR(32), storing as stringified hex values.
+
+    from SQLAlchemy Docs
+    http://docs.sqlalchemy.org/en/rel_0_9/core/custom_types.html
+    """
+    impl = CHAR
+
+    def load_dialect_impl(self, dialect):
+        if dialect.name == 'postgresql':
+            return dialect.type_descriptor(UUID())
+        else:
+            return dialect.type_descriptor(CHAR(32))
+
+    def process_bind_param(self, value, dialect):
+        if value is None:
+            return value
+        elif dialect.name == 'postgresql':
+            return str(value)
+        else:
+            if not isinstance(value, uuid.UUID):
+                return "%.32x" % uuid.UUID(value).int  # noqa # pylint: disable=no-member
+            else:
+                # hexstring
+                return "%.32x" % value.int
+
+    def process_result_value(self, value, dialect):
+        if value is None:
+            return value
+        else:
+            return uuid.UUID(value)
