+++ tox.ini
@@ -20,6 +20,6 @@ deps =
 commands =
     pip install -U pip
     python setup.py install
-    py.test --basetemp={envtmpdir}
+    py.test --basetemp={envtmpdir} tests python_model_service/orm/test.py
